Move parent/gamer dashboards to single-page scroll with section pills
Parent and gamer roles drop the sidebar and gain a Home-style pill nav.
Parent: My Family (placeholder) / Feedback / Settings.
Gamer:  My Family (placeholder) / Yty / Feedback / Settings.

Settings and Feedback content is extracted into shared section
components so the standalone /settings and /feedback routes (still used
by admin/gedu) and the new in-page dashboard sections render the same
thing. Admin/gedu sidebar experience is unchanged.

Orphaned sub-routes (/parent/gamers/*, /parent/groups/[id],
/gamer/groups/*, /gamer/voice/[id]) and dead sidebar config for
customer/gamer roles are documented in TODO.md for follow-up cleanup.
