+++ TODO.md
@@ -2,6 +2,17 @@
 
 ## Cleanup
 
+- [ ] **Delete orphaned parent/gamer sub-routes left behind by the unified-navbar single-page-scroll redesign.** When the parent and gamer dashboards moved to the in-page section-pill UX (My Family / [Yty] / Feedback / Settings on `/parent` and `/gamer`), the sidebar and the dashboard quick-action cards that linked to these routes were removed. The routes themselves still resolve, but no in-app navigation points at them anymore — they're reachable only by typing the URL or following an external/email link. Replace each one (or its functionality) inside the new dashboard sections, then delete:
+  - `src/app/(dashboard)/parent/gamers/page.tsx` and `src/app/(dashboard)/parent/gamers/[id]/page.tsx` — to be replaced by real content inside the parent's `#my-family` section.
+  - `src/app/(dashboard)/parent/groups/[id]/page.tsx` — parent group-detail view; entry point gone.
+  - `src/app/(dashboard)/gamer/groups/page.tsx` and `src/app/(dashboard)/gamer/groups/[id]/page.tsx` — to be replaced by content inside the gamer's `#my-family` (or a new groups section). The voice chat at `src/app/(dashboard)/gamer/voice/[id]/page.tsx` is currently entered via the groups → group-detail chain, so do NOT delete `voice/[id]` until a new path into a live session exists.
+  - `customer` and `gamer` entries in `navItemsByRole` in `src/components/layout/sidebar.tsx` — dead config now that the sidebar no longer renders for those roles. Safe to delete after the routes above are gone.
+  - Sidebar-only translation keys that exclusively serve those entries (`sidebar.myGamers`, `sidebar.home`, `sidebar.myGroups` in en/fi/sv/tlh) once the config above is dropped.
+- [ ] **Decide what to do with `/settings` and `/feedback` for parent/gamer roles.** The shared `<SettingsSectionContent />` and `<FeedbackSectionContent />` components are now rendered inside the parent/gamer dashboards as `#settings` and `#feedback` scroll sections. The standalone routes still exist (admin and gedu link to them from their sidebars). The intended end state — per the user's plan — is that parent/gamer have only `/parent#settings` / `/gamer#feedback` etc., and the standalone routes either redirect or stop existing for those roles. Two open subtasks:
+  - For parent/gamer hitting `/settings` or `/feedback` directly (bookmark/old link), add a server-side redirect to `/parent#settings`, `/gamer#feedback`, etc. Probably in the route's `page.tsx` via `getUserWithProfile()` + `redirect(...)`.
+  - Decide whether admin and gedu eventually adopt the same in-page-scroll dashboard pattern. If yes, the shared section components are already extracted — only the page composition would need to follow.
+- [ ] **Add hash-anchor deep-link support to parent/gamer dashboards** (the same trick `/admin/ui-components` uses). The pill currently smooth-scrolls in response to clicks but does not honour the hash on initial load — landing on `/parent#feedback` from an external link won't auto-scroll to the feedback section. Out of scope when the unified navbar shipped, easy to add later.
+- [ ] **Sub-routes under `/parent/*` and `/gamer/*` have no back-affordance now that the sidebar is gone.** Reaching `/parent/gamers/[id]`, `/gamer/groups/[id]`, `/gamer/voice/[id]`, etc. leaves the user with only the header avatar to return to the dashboard root — no breadcrumb, no back link. Acceptable transient state because these routes are slated for deletion (see first item), but if any of them outlive the cleanup, add an in-page back link.
 - [ ] Remove `FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true` from `.github/workflows/ci.yml` after June 2, 2026 (Node.js 24 becomes the default runner)
 - [ ] Add CHECK constraints to `profiles.locale` (`IN ('en', 'fi', 'sv', 'tlh')`) and `profiles.currency` (`IN ('EUR', 'SEK', 'USD', 'GBP')`) — both are plain text columns with app-level validation only
 - [ ] **Eliminate the extra reload on first-device-login for locale-dependent SSR.** `LocaleProvider` currently reconciles a stale `locale` cookie with `profile.locale` on mount and calls `router.refresh()`, which produces a visible second render on the first page after signing in on a new device. The underlying issue: next-intl's `getRequestConfig` runs before auth is available and can only read cookies/headers, so a client-side reconciliation step is the only option. This is a canary — any future pre-render preference (timezone, theme-critical CSS, feature flags, anything that needs to be baked into SSR output) will hit the same pattern where "cookie is a cache of profile state" and new devices always miss the cache. Two architectural fixes worth considering: (a) write preference cookies server-side during the auth callback so SSR sees them on the very next request (cheap per login, no per-render cost), or (b) thread the authenticated user through `getRequestConfig` so it can read the profile directly (per-request DB cost but no divergence possible). (a) is the better default.
