+++ messages/tlh.json
@@ -1011,6 +1011,13 @@
     "failedToSubmit": "QIn ngeH Dabe'lu'.",
     "failedToSubmitRetry": "QIn ngeH Dabe'lu'. yIboq."
   },
+  "dashboardSections": {
+    "myFamily": "qorDu'wIj",
+    "yty": "Yty",
+    "feedback": "QaH",
+    "settings": "tlhInganpu' chu'",
+    "myFamilyDescription": "Daq vam chenmoHlu'taH. tugh qorDu'lIj naQ Dachelqu' wa' Daq."
+  },
   "footer": {
     "tagline": "toy'wI' ghom - DaH bIQam, batlh yIghoS.",
     "copyright": "© {year} toy'wI' ghom. Hoch batlh.",
