+++ src/app/(dashboard)/parent/page.tsx
@@ -1,61 +1,60 @@
 import type { Metadata } from "next";
-import Link from "next/link";
 import { useTranslations } from "next-intl";
 import { getTranslations } from "next-intl/server";
-import { Gamepad2, ShoppingCart, Settings } from "lucide-react";
-import { Card, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { ROUTES } from "@/lib/constants";
+import { Users } from "lucide-react";
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
+import { DashboardSectionPill, type DashboardSection } from "@/components/layout";
+import { FeedbackSectionContent } from "@/components/feedback/feedback-section-content";
+import { SettingsSectionContent } from "@/components/settings/settings-section-content";
 
 export async function generateMetadata(): Promise<Metadata> {
   const t = await getTranslations("metadata.pages");
   return { title: t("parentDashboard"), description: "Manage your gamers and enrollments" };
 }
 
-const quickActionIcons = [Gamepad2, ShoppingCart, Settings] as const;
-const quickActionHrefs = [
-  ROUTES.customer.gamers,
-  ROUTES.products,
-  ROUTES.settings,
-] as const;
-const quickActionKeys = ["gamers", "browseClubs", "settings"] as const;
-
 export default function CustomerDashboardPage() {
-  const t = useTranslations('parent');
+  const t = useTranslations('dashboardSections');
+
+  const sections: DashboardSection[] = [
+    { id: 'my-family', label: t('myFamily') },
+    { id: 'feedback', label: t('feedback') },
+    { id: 'settings', label: t('settings') },
+  ];
 
   return (
-    <div className="space-y-6">
-      <div>
-        <h1 className="text-3xl font-bold">{t('dashboard.welcome')}</h1>
-        <p className="text-muted-foreground">
-          {t('dashboard.subtitle')}
-        </p>
-      </div>
+    <>
+      <DashboardSectionPill sections={sections} ariaLabel={t('myFamily')} />
 
-      <div className="grid gap-4 md:grid-cols-2">
-        {quickActionKeys.map((key, i) => {
-          const Icon = quickActionIcons[i];
-          const href = quickActionHrefs[i];
-          const secondary = i >= 1;
-          return (
-          <Link key={href} href={href}>
-            <Card className="group h-full transition-colors hover:bg-accent hover:text-accent-foreground">
+      <div className="space-y-24 pb-24">
+        <section id="my-family" className="scroll-mt-32">
+          <div className="mx-auto max-w-3xl space-y-6">
+            <h1 className="text-3xl font-bold">{t('myFamily')}</h1>
+            <Card>
               <CardHeader>
                 <div className="flex items-center gap-4">
-                  <div className={`flex h-12 w-12 items-center justify-center rounded-lg ${secondary ? "bg-secondary/10" : "bg-primary/10"}`}>
-                    <Icon className={`h-6 w-6 ${secondary ? "text-secondary" : "text-primary"}`} />
+                  <div className="flex h-12 w-12 items-center justify-center rounded-lg bg-primary/10">
+                    <Users className="h-6 w-6 text-primary" />
                   </div>
-                  <div>
-                    <CardTitle className="text-lg">{t(`dashboard.quickActions.${key}.title`)}</CardTitle>
-                    <CardDescription>{t(`dashboard.quickActions.${key}.description`)}</CardDescription>
-                  </div>
+                  <CardTitle className="text-lg">{t('myFamily')}</CardTitle>
                 </div>
               </CardHeader>
+              <CardContent>
+                <CardDescription className="text-base">
+                  {t('myFamilyDescription')}
+                </CardDescription>
+              </CardContent>
             </Card>
-          </Link>
-          );
-        })}
-      </div>
+          </div>
+        </section>
 
-    </div>
+        <section id="feedback" className="scroll-mt-32">
+          <FeedbackSectionContent />
+        </section>
+
+        <section id="settings" className="scroll-mt-32">
+          <SettingsSectionContent />
+        </section>
+      </div>
+    </>
   );
 }
