+++ src/components/layout/index.ts
@@ -3,4 +3,6 @@ export { Sidebar } from "./sidebar";
 export { Footer } from "./footer";
 export { Copyright } from "./copyright";
 export { DashboardLayout } from "./dashboard-layout";
+export { DashboardSectionPill } from "./dashboard-section-pill";
+export type { DashboardSection } from "./dashboard-section-pill";
 export { MouseflowConsent } from "./mouseflow-consent";
